+++ packages/utils/src/index.ts
@@ -4,6 +4,7 @@
  */
 
 export * from './formatters';
+export * from './data-local';
 export * from './fechamento';
 export * from './lucro';
 export * from './despesa';